requirements.txt
# Python Dependencies for Big Data Processing Pipeline

# Core Airflow
apache-airflow==2.7.3
apache-airflow-providers-amazon==8.11.0

# AWS SDK
boto3==1.34.14
botocore==1.34.14

# Data Processing
pyspark==3.5.0
pandas==2.1.4
numpy==1.26.2

# Configuration Management
pyyaml==6.0.1
python-dotenv==1.0.0

# Data Validation
great-expectations==0.18.8

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and Monitoring
watchtower==3.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
moto==4.2.11  # AWS mocking

# Development
black==23.12.1
flake8==7.0.0
pylint==3.0.3

# Optional: Database drivers if using for metadata
psycopg2-binary==2.9.9  # PostgreSQL
pymysql==1.1.0  # MySQL

# Optional: Data visualization
matplotlib==3.8.2
seaborn==0.13.0
